Increase the --max-peers count on all fleet nodes
This setting is recommended due to the high-number of validators
attached to each node which requires maintaining healthy peer
counts on a larger number of gossip topics. For now, we don't
bother specifying a lower value on the servers running a lower
number of validators in order to simplify the implementation.