nimbus.ropsten: lower TTD to 50000000000000000
"Here’s an update on the Ropsten TTD: we’ve chosen to use a value of
50000000000000000. At the current hash rate, it would hit in several
weeks. That said,  the EF will rent hash rate itself and work with
partners to accelerate things. With a modest increase in hash rate, we
should be able to make the Ropsten Merge happen on June 8/9.  If your
team is running any validators, please perform a TTD override to
50000000000000000 on them ASAP."

Signed-off-by: Jakub Sokołowski <jakub@status.im>