holesky,hoodi: add monitor start delay based on val count
This is a hack-fix for Consul ignoring critical threshold setting when
it's reloaded or restarted, which is triggered by the monitor script.

If we delay monitor script run we can avoid Consul service going
straight to critical state due to the unavailable REST and merics
endpoint while a large number of validators is being loaded.

For more info see:
https://github.com/status-im/infra-hq/issues/235

Signed-off-by: Jakub Sokołowski <jakub@status.im>